beast_resnet port
