+++ beast/models/beast_resnet/beast_resnet_train.py
@@ -1,0 +1,9 @@
+"""Training entry point for the ResNet autoencoder.
+
+Delegates to the shared beast.train.train function, which handles
+BaseDataset / BaseDataModule and epoch-based training.
+"""
+
+from beast.train import train
+
+__all__ = ['train']
