+++ tests/test_config.py
@@ -8,11 +8,11 @@ from pydantic import ValidationError
 from beast.config import (
     BeastConfig,
     OptimizerConfig,
-    ResnetModelParams,
     TrainingConfig,
     VitModelParams,
 )
 from beast.io import load_config
+from beast.models.beast_resnet.beast_resnet_config import ResnetModelParams
 
 _CONFIGS_DIR = Path(__file__).parent.parent / 'configs'
 _CONFIG_FILES = list(_CONFIGS_DIR.glob('*.yaml'))
